Si el fichero existe entonces no lo subo.

+++ upload.php
@@ -8,6 +8,8 @@ if (!empty($_FILES)) {
     $tempFile = $_FILES['file']['tmp_name'];
     $targetPath = dirname( __FILE__ ) . $ds. $storeFolder . $ds;
     $targetFile =  $targetPath. $_FILES['file']['name'];
-    move_uploaded_file($tempFile,$targetFile);
+    if (!file_exists($targetFile)) {
+        move_uploaded_file($tempFile,$targetFile);
+    }
 }
 ?>
